Updated documentation
I created all of the organisational files and filled in the rules and planning of the program. I have also attached my excalidraw save that contains my info for ease of acess. No changes to my code as purely documentational updates. Next I need to continue placing the rules in a pretty format, do the requirements and specifications and do the UI and usecase planning