Major documentation update
I added and completed all of the diagrams.  The requirements and specifications are done. No changes to my code as purely documentational updates. Now working hard on the prototype, I have decided to do the GUI mockup instead of the functional text based version. May need to slightly edit some of the diagrams before evaluating the sprint 1.