Began creation of the game logic.
The work was partly completed on the weekend. I am not finished with the design side but need to begin coding to clarify how my flowcharts and psudeocode will work.  It's essential I get this stage correct as these are the fundamental's of my game. I have made drawing and card creation functions, I need to continue work on my game logic to continue with the design phase. I plan to create some of the basic card functions into OOP for simplicity.