Implementation of OOP practices
I began implementation of Classes to clean up my game logic code. I have created enough of an understanding to be able to continue on the design phase of my project.  I have also made a change to my game to make the program and logic easier to code and adds an element of intrigue and surprise to my game. The OOP is essential in improving my codes readability and structure. I will need to continue my development of the program and begin integrating.  The programming of integration will be continued after I finalize my diagrams for the design aspect of the project now that I have created a clear understanding.