Update to game logic and main+ Documentation
I have used percentage of screen to convert most of the game into measurements that all resolutions can use. I have not completed the game flow chart or structure chart as am a little lost on the code and am wishing clarification in class. I have however begun integrating and as my program doesn't require much of a third phase besides the theory its ok if it runs slightly overtime. It's crucial I get the fundamentals right now to have a functional and fun game. Next I need to finish my code in order to complete the game flowchart and structure chart to then complete the review for this sprint.