Update to code and game logic.
I have decided to completely change the rules of my game and make players go at the same time for a more fun challenge. I will now adjust all of the game logic to follow this. The commit I'm making is a save point as such. The next commit will be a gigantic change to my project, rules and documentation. I plan to rework my entire main functions and complete sprint 2 to move onto debugging and theory for sprint 3.