Big update to code docs and game logic .py+ main.py
I have completed the second sprint, the rest of the code development will be done in the 4th sprint (maybe 3rd), I need advice from the teacher to clarify my concerns around structure charts. I have completely redone a lot of the games structure and processing to make it all work. I have also added in updated flowcharts and pseudocode.  The rework of my game's structure has significantly changed the end result and outcome for my game and the working input system was quite diffrent to my original vision. All that is left for sprint 2 is the review and clarification of ideas about the structure chart, I will then do the class diagram and complete the rest of the theory and code to finish off the project! I have also packaged an exe for simplicity!